Renamed CommandRunner to ProcessInteraction
