+++ de.dlr.sc.virsat.commons/META-INF/MANIFEST.MF
@@ -12,9 +12,9 @@ Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Bundle-ActivationPolicy: lazy
 Eclipse-ExtensibleAPI: true
 Export-Package: de.dlr.sc.virsat.commons,
- de.dlr.sc.virsat.commons.command,
  de.dlr.sc.virsat.commons.datastructures,
  de.dlr.sc.virsat.commons.debug,
  de.dlr.sc.virsat.commons.exception,
+ de.dlr.sc.virsat.commons.external,
  de.dlr.sc.virsat.commons.file
 Automatic-Module-Name: de.dlr.sc.virsat.commons
